requirements.txt
asgiref
atomicwrites
attrs
autopep8
django-adminactions
black
certifi
charset-normalizer
click
colorama
dj-database-url
Django
django-compat
django-crispy-forms
django-heroku
django-js-asset
django-markdown-deux
django-model-utils
django-online-users
django-redis
django-simple-history
django-taggit
djangorestframework
gunicorn
humanize
idna
importlib-metadata
iniconfig
Markdown
markdown2
martor
mypy-extensions
mysqlclient
packaging
pathspec
Pillow
platformdirs
pluggy
portalocker
psycopg2
py
pycodestyle
pyparsing
pytest
python-slugify
pytz
redis
regex
requests
six
sqlparse
text-unidecode
toml
tomli
typed-ast
typing-extensions
tzlocal
urllib3
whitenoise
xlrd
xlwt
yapf
zipp